The update is now working on save
However the interactions with file versions is currently broken and need something
