+++ frontend/src/components/visualizations/JSONVisualizer/manifest.json
@@ -2,7 +2,7 @@
 	"name": "json-editor",
 	"version": "1.0.0",
 	"description": "A React component for visualizing and editing JSON data",
-	"main": "JSON.tsx",
+	"main": "JSONVisualizer.tsx",
 	"dependencies": {
 		"clowder2-core": "1.0.0",
 		"react": "^17.0.2",
@@ -11,7 +11,7 @@
 		"@codemirror/lang-json": "^6.0.1"
 	},
 	"visConfig": {
-		"name": "JSON",
+		"name": "JSONVisualizer",
 		"mainType": "application",
 		"mimeTypes": ["application/json"],
 		"fields": []
